Moved util to utils
