+++ aiapy/calibrate/tests/test_meta.py
@@ -8,8 +8,8 @@ from astropy.table import QTable
 from astropy.time import Time, TimeDelta
 
 from aiapy.calibrate import update_pointing
-from aiapy.calibrate.util import get_pointing_table
-from aiapy.util.exceptions import AIApyUserWarning
+from aiapy.calibrate.utils import get_pointing_table
+from aiapy.utils.exceptions import AIApyUserWarning
 
 
 @pytest.fixture
