+++ aiapy/response/tests/test_channel.py
@@ -8,7 +8,7 @@ import astropy.units as u
 
 from sunpy.util.metadata import MetaDict
 
-from aiapy.calibrate.util import get_correction_table
+from aiapy.calibrate.utils import get_correction_table
 from aiapy.response import Channel
 from aiapy.response.channel import VERSION_NUMBER
 from aiapy.tests.data import get_test_filepath
