+++ aiapy/tests/test_idl.py
@@ -10,7 +10,7 @@ import pytest
 import astropy.units as u
 
 from aiapy.calibrate import estimate_error
-from aiapy.calibrate.util import get_error_table
+from aiapy.calibrate.utils import get_error_table
 from aiapy.conftest import CHANNELS
