+++ aiapy/utils/__init__.py
@@ -5,4 +5,4 @@ Subpackage with miscellaneous utility functions.
 from .decorators import *
 from .exceptions import *
 from .net import *
-from .util import *
+from .utils import *
