+++ aiapy/utils/exceptions.py
