+++ aiapy/utils/tests/test_utils.py
@@ -2,14 +2,14 @@ import pytest
 
 from astropy.tests.helper import assert_quantity_allclose
 
-import aiapy.util
-from aiapy.util.util import _QUALITY_FLAG_MESSAGES
+import aiapy.utils
+from aiapy.utils.utils import _QUALITY_FLAG_MESSAGES
 
 
 @pytest.mark.remote_data
 def test_sdo_location(aia_171_map) -> None:
     # Confirm that the queried location matches AIAMap's interpretation of the FITS file
-    result = aiapy.util.sdo_location(aia_171_map.date)
+    result = aiapy.utils.sdo_location(aia_171_map.date)
     aia_171_map.observer_coordinate.transform_to(result)
     assert_quantity_allclose(result.cartesian.xyz, result.cartesian.xyz)
 
@@ -18,7 +18,7 @@ def test_sdo_location(aia_171_map) -> None:
 def test_sdo_location_raises_error() -> None:
     # Confirm that an error is raised for a time without records
     with pytest.raises(RuntimeError, match="No data found for this query"):
-        aiapy.util.sdo_location("2001-01-01")
+        aiapy.utils.sdo_location("2001-01-01")
 
 
 @pytest.mark.parametrize(
@@ -37,4 +37,4 @@ def test_check_quality_flag(bits):
     messages = ["nominal"]
     if bits:
         messages = [_QUALITY_FLAG_MESSAGES.get(b, "(empty)") for b in bits]
-    assert messages == aiapy.util.check_quality_flag(quality)
+    assert messages == aiapy.utils.check_quality_flag(quality)
