+++ changelog/366.breaking.rst
@@ -1,0 +1,1 @@
+Renamed all "util" modules to "utils".
