--- docs/reference/util.rst
@@ -1,7 +1,0 @@
-==========================
-Utilities (``aiapy.util``)
-==========================
-
-.. automodapi:: aiapy.util
-    :skip: AIApyWarning, AIApyUserWarning
-    :no-inheritance-diagram:
