+++ docs/reference/utils.rst
@@ -1,0 +1,7 @@
+===========================
+Utilities (``aiapy.utils``)
+===========================
+
+.. automodapi:: aiapy.utils
+    :skip: AIApyWarning, AIApyUserWarning
+    :no-inheritance-diagram:
