+++ examples/calculate_response_function.py
@@ -14,7 +14,7 @@ import matplotlib.pyplot as plt
 import astropy.time
 import astropy.units as u
 
-from aiapy.calibrate.util import get_correction_table
+from aiapy.calibrate.utils import get_correction_table
 from aiapy.response import Channel
 
 ###############################################################################
