+++ examples/download_specific_data.py
@@ -20,7 +20,7 @@ import sunpy.map
 from sunpy.net import Fido, attrs
 
 from aiapy.calibrate import correct_degradation, register, update_pointing
-from aiapy.calibrate.util import get_correction_table, get_pointing_table
+from aiapy.calibrate.utils import get_correction_table, get_pointing_table
 
 ###############################################################################
 # Exporting data from the JSOC requires registering your
